x64 generator: Fix store asm, make inline asm generation more ergonomic
lf_op_store was using incorrect operands.

I can now write the asm for each operation in a constant string that is
then transformed into the correct asm for the type's size. This makes it
much easier to read and modify the asm. It isn't very complex, but here
are the differences in my asm template:
- Apostrophes (') are replaced with quotes ("). This prevents me from
  writing the asm with \" everywhere, which makes it harder to read.
- Instructions are replaced with the same instruction with the size
  suffix. For example, mov is replaced with movb if the type is of size
  1. This means instructions are reserved words and should not be used
  for anything else.