x64 generator: Add Double Width CAS and CASX (dcas, dcasx) for x64
Implement DCAS for all supported types on x64. For types smaller than
8 bytes, we use the underlying cas or casx operation for a type double
the size. This doesn't break strict aliasing rules (at least by C11's
standards) because we use a union with a scalar of the larger type and
tuple of the smaller type. For types that are 8 bytes, we use
cmpxchg16b.

dcasx takes a bool * as a param unlike regular casx. The motivation
behind this was that checking if casx succeeded is a simple equality
check. Checking dcasx is a little more annoying because you have to
check each item in the tuple. This can be annoying and getting the ZF
flag after a CAS is essentially free. I may change this before 1.0 if I
find myself not using the success flag.