x64 generator: Add option to emit support macros
Not all architectures will support all operations on all types. For
example, there is no way to implement a lock free uint64_t DCAS on
ARM32. This commit addds additional functionality to the x64 generator
which emits macros for each type and function to indicate support. This
becomes important if the user is targeting multiple architectures
(especially 32 bit ones that may not be able to implement everything).