gcc generator: Make CAS strong
CAS being strong by default (not failing spuriously) is what I want.
Much of the code I write assumes that if a CAS fails someone else was
able to do it. This assumption is important in the context of managing
"one time" operations like freeing a resource. Imagine I allocate a new
buffer, copy the old one, and modify it. I then want to CAS the global
pointer to this buffer and free the old one. If my CAS fails, it should
mean someone else is freeing the buffer. If the CAS fails spuriously,
then the resource leaks.