lf_ring: Use lf_op_casx for swapping head/tail in mp/mc
lf_ring was using lf_op_cas for swapping the head/tail. This is fine but
it guarantees that we issue a lf_op_load at the start of every loop. If
we use lf_op_casx instead, we can load a lf_op_load before every loop
and let casx give us the value on failure. On x64, this is much more
efficient because it only requires a mov from the a register (the
compiler could likely optimize this by keeping head/tail in the a register).